v0.0.1.1 - 25.03.2021.
Having a lot of trouble with save games. Had to revert to previous branch 4-5 different times because it didn't work. I am back to designing the pipeline again, since I can't get it to work as I want (or as close as I want). Also had problems, not relating to the game, so didn't update during last week.

Did:

1) Removed "continue playing" widget. Save, and load, game are all located in "New Game".
2) Added 3 save game slots - player can play 3 different games.
3) Save game - only current level is stored.
4) Implemented Game instance - that will allow me to transfer scores, levels, etc. across different game modes.
5) Implemented pause menu - player can pause the game while playing the game 'Esc'. It acts as a main menu, except it has additional "Restart Level" and "Quit" options.
  5.1) Restart Level - restarts current level, also possible with 'r' key.
  5.2) Quit - Returns player to main menu screen. Exit will leave the game.

All in all, not really pleased by this update. Have to do a lot of improvements on the save game. I want to finish those this week.