v0.0.1.3 - 01.04.2021.
With this, I can say that my save/load/delete feature is complete. Again, most of these changes were in BP's. Most of interactions (for this feature) are located in Widgets, really don't want to code widgets in C++ (I can make C++ base class and BP derived class so I can style it). Might do these in the future, it's not difficult, just annoying to deal with.

1) Fixed save games path - Save games weren't loading on beginning of the new session. Apparently 'FPaths::ProjectSavedDir()' doesn't actually go inside 'SaveGames' dir. It was my bad for thinking that it does, without testing it afterwards (was getting late and just wanted to finish that part of the feature).
2) Fixed cancel save game - Now clicking on the 'return' button, or pressing the 'esc' key on keyboard, won't save the game no matter what the player types in.
3) Fixed NameSaveSlot_WBP focus - On player click, with mouse, widget was deleted, and gave save slot was saved based on current input. Now it will only change focus to widget instead of editable text box.
4) Reworked NewGameSlot_WBP - Now it holds save/load slot and delete slot. GUI depends if there exists, or doesn't, a save game.
5) Implemented 'esc' key - In menus pressing 'esc' key on the keyboard will return player to the previous menu or cancel the current action (exit, naming the game). Focus will always be on the current menu (this was a pain to understand).
6) Implemented delete feature - Player can now delete his saved game. Delete button will only appear if player has made a save game. Otherwise it will not be shown (does exist always).
7) Implemented enter feature - Player can save game with pressing 'enter' on  the keyboard. No longer need to click on 'save' with the left mouse button.
8) Implemented save game limitations - Things like naming and length. This was fun to do (not).
  8.1) Save game name limitation - Player can only use standard letters (a-q/A-Q) and numbers (0-9).
  8.2) Save game name length limit - Player can use 12 characters for naming a save game. Yes, save slot can be empty and it will work. Why? Why not.

Time for the next challenge. Much less coding in this one, probably only for some 'feature' fixes (when I stumble upon them). Time for some modeling/remodeling, texturing, animations, particle effects (this one is second to last, never did this) and sound effects (this one is last, last time I did this was in college).